requirements.txt
llama_index
llama-index-vector-stores-faiss
langchain
openai
faiss-cpu
pypdf
python-dotenv
streamlit
azure.cognitiveservices.speech
SpeechRecognition 
pyaudio